requirements.txt
boto3==1.39.11
botocore==1.39.11
duckdb==1.3.2
jmespath==1.0.1
loguru==0.7.3
numpy==2.3.1
pandas==2.3.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
s3transfer==0.13.1
six==1.17.0
tzdata==2025.2
urllib3==2.5.0
